Refactored hky to separate module
